Added Stricted Authorization

+++ backend/src/main/java/org/example/backend/forum/ForumService.java
@@ -2,7 +2,9 @@ package org.example.backend.forum;
 
 import lombok.RequiredArgsConstructor;
 import org.bson.types.ObjectId;
+import org.example.backend.deletion.AccessService;
 import org.example.backend.deletion.CascadeDeletionService;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
 import java.time.Instant;
@@ -13,6 +15,7 @@ public class ForumService {
 
     private final ForumRepository forumRepository;
     private final CascadeDeletionService deletionService;
+    private final AccessService accessService;
 
     public Forum createForum(ForumCreationRequest request, Long userId) {
 
@@ -26,9 +29,14 @@ public class ForumService {
         return forumRepository.save(forum);
     }
 
-    public void deleteForum(ObjectId forumId) {
+    public void deleteForum(ObjectId forumId, long userId) {
+        if (!accessService.canDeleteForum(longToObjectId(userId), forumId)) {
+            throw new AccessDeniedException("User " + " cannot delete this forum");
+        }
+
         deletionService.deleteForum(forumId);
     }
+
 
 
     private ObjectId longToObjectId(Long value) {
